Added $currency to the content items.
It's an arguable change because we have currency in the event metadata (`event_data.currency`) for *commerce* events, but no indication of currency in the *content* events.

Currently, the content items only have a price with no indication of currency. For consistency of the content items across all events and since we have this field open in EO anyways, I suggest we add it to the docs.